Fixed permission test failures
